Refactor undo stack into separate module
